[implemented] created implementation of cryptocurrency repository interface

+++ app/src/main/java/com/jvmori/cryptocurrencyapp/cryptolist/domain/entities/CryptocurrencyEntity.kt
@@ -3,8 +3,8 @@ package com.jvmori.cryptocurrencyapp.cryptolist.domain.entities
 data class CryptocurrencyEntity(
     var cryptoName: String,
     var symbol: String,
-    var priceUSD: Float,
-    var percentChangeHour: Float,
-    var percentChange24Hours: Float,
-    var volume: Float
+    var priceUSD: Double,
+    var percentChangeHour: Double,
+    var percentChange24Hours: Double,
+    var volume: Double
 )
